Add "Merge Responsibilities" document
Adds a document sketching out the responsibilities of a merger/deployer.
Mostly a skeleton to be fleshed out as we develop the outlines of our
auto-deploy process.

Also adds a few extra points to the code review document which came up
during discussion of merge responsibilities, and one more suggested
reading citation that had some good & relevant points.